requirements.txt
fastapi
uvicorn[standard]
tensorflow
opencv-python-headless
numpy
python-multipart
aiofiles
mediapipe
google.generativeai
gtts